// src/interfaces/FetchPokemonInterfaces.ts
export interface Pokemon {
   id: number;
   name: string;
   picture: string | undefined;
   type: string;
   url: string;
   ability: string;
   height: number;
   weight: number;
   capture_rate: number;
   description: string;
   generation: number;
}